fix: refactor the whole code

+++ src/header.h
@@ -17,9 +17,9 @@ struct Record
     int userId;
     char name[100];
     char country[100];
-    int phone;
+    long long phone;
     char accountType[10];
-    int accountNbr;
+    long long accountNbr;
     double amount;
     struct Date deposit;
     struct Date withdraw;
